refactor i18n translation
