+++ packages/main-ui/src/pages/politician/status/i18n.rs
@@ -3,6 +3,11 @@ use dioxus_translate::translate;
 translate! {
     PoliticianStanceTranslate;
 
+    title: {
+        ko: "국회의원",
+        en: "Politician",
+    },
+
     supportive: {
         ko: "찬성",
         en: "Supportive",
@@ -26,5 +31,40 @@ translate! {
     more: {
         ko: "더보기",
         en: "MORE",
+    }
+
+    name: {
+        ko: "이름",
+        en: "NAME",
+    }
+
+    party: {
+        ko: "정당",
+        en: "PARTY",
+    }
+
+    district: {
+        ko: "지역구",
+        en: "DISTRICT",
+    }
+
+    stance_on_crypto: {
+        ko: "암호화폐에 대한 입장",
+        en: "STANCE ON CRTPTO",
+    }
+
+    proclaim: {
+        ko: "PROCLAIM",
+        en: "PROCLAIM",
+    }
+
+    tooltip: {
+        ko: "Proclaim 은 해당 의원의 암호화폐에 대한 정책적 긍정 및 부정에 대한 의사를 표현하는 것입니다. 해당 의원실 소속원만 변경할 수 있습니다.",
+        en: "Proclaim is a way to express the policy of a politician on cryptocurrency. Only the staff of the politician can change it.",
+    }
+
+    change_stance: {
+        ko: "입장 변경",
+        en: "CHANGE STANCE",
     }
 }
